Bug fixed for an exception generated when you don't have monitored folders registered in the database.

+++ Data/FilesHandler.cs
@@ -45,34 +45,38 @@ namespace Data
 
         public void OrganizeFiles(string action,string destination)
         {
-            switch (action)
+            //This condition will prevent an exception if there are no monitored folders.
+            if (this.directoryFilesInfo != null)
             {
-                case "Move":
-                    for (int i = 0; i < this.directoryFilesInfo.Length; i++)
-                    {
-                        Match match = Regex.Match(directoryFilesInfo[i], @"([a-zA-Z0-9-_])+\.[a-zA.Z0-9]+");
-                        if (match.Success)
+                switch (action)
+                {
+                    case "Move":
+                        for (int i = 0; i < this.directoryFilesInfo.Length; i++)
                         {
-                            string FileName = "\\" + match.Value;
+                            Match match = Regex.Match(directoryFilesInfo[i], @"([a-zA-Z0-9-_])+\.[a-zA.Z0-9]+");
+                            if (match.Success)
+                            {
+                                string FileName = "\\" + match.Value;
+                                try
+                                {
+                                    File.Move(directoryFilesInfo[i], destination + FileName);
+                                }
+                                catch (Exception) { }
+                            }
+                        }
+                        break;
+
+                    case "Delete":
+                        for (int i = 0; i < this.directoryFilesInfo.Length; i++)
+                        {
                             try
                             {
-                                File.Move(directoryFilesInfo[i], destination + FileName);
+                                File.Delete(directoryFilesInfo[i]);
                             }
                             catch (Exception) { }
                         }
-                    }
-                break;
-
-                case "Delete":
-                    for (int i = 0; i < this.directoryFilesInfo.Length; i++)
-                    {
-                        try
-                        {
-                            File.Delete(directoryFilesInfo[i]);
-                        }
-                        catch (Exception) { }
-                    }
-                break;
+                        break;
+                }
             }
         }
